Moved the userlogin content into the correct folder structure.
