+++ app/src/main/java/com/example/notes2text/adapters/AccountSettingsAdaptors/PopUpEmailController.java
@@ -6,8 +6,8 @@ import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.notes2text.UserLogin.UserRep;
-import com.example.notes2text.UserLogin.UserUpdateInfo;
+import com.example.notes2text.ui.userloginframeworks.UserRep;
+import com.example.notes2text.usecases.userloginusecases.UserUpdateInfo;
 
 /**
  * Controller class for changing the email.
