+++ app/src/main/java/com/example/notes2text/adapters/userloginadaptors/RegisterView.java
@@ -1,4 +1,4 @@
-package com.example.notes2text.UserLogin;
+package com.example.notes2text.adapters.userloginadaptors;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -10,9 +10,7 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.notes2text.R;
-
-import java.util.Arrays;
-import java.util.List;
+import com.example.notes2text.usecases.userloginusecases.RegisterUseCase;
 
 public class RegisterView extends AppCompatActivity {
