+++ app/src/main/java/com/example/notes2text/ui/userloginframeworks/UserRep.java
@@ -1,4 +1,4 @@
-package com.example.notes2text.UserLogin;
+package com.example.notes2text.ui.userloginframeworks;
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -6,8 +6,7 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import java.util.HashMap;
-import java.util.Map;
+import com.example.notes2text.usecases.userloginusecases.UserRepository;
 
 public class UserRep extends SQLiteOpenHelper implements UserRepository {
