+++ app/src/main/res/layout/activity_main2.xml
@@ -5,7 +5,7 @@
     android:id="@+id/login"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserLogin.RegisterView">
+    tools:context=".adapters.userloginadaptors.RegisterView">
 
     <TextView
         android:id="@+id/register"
