+++ app/src/test/java/com/example/notes2text/UserLogin/UserRequirementsTest.java
@@ -2,6 +2,8 @@ package com.example.notes2text.UserLogin;
 
 import static org.junit.Assert.*;
 
+import com.example.notes2text.usecases.userloginusecases.UserRequirements;
+
 import org.junit.Test;
 
 /**
@@ -18,17 +20,17 @@ public class UserRequirementsTest {
     public void validPassword() {
         UserRequirements requirements = new UserRequirements();
         String password = "Bill1$";
-        assertEquals(true, requirements.validPassword(password));
+        assertTrue(requirements.validPassword(password));
         password = "bill1$";
-        assertEquals(false, requirements.validPassword(password));
+        assertFalse(requirements.validPassword(password));
         password = "BILL1$";
-        assertEquals(false, requirements.validPassword(password));
+        assertFalse(requirements.validPassword(password));
         password = "Bill 1$";
-        assertEquals(false, requirements.validPassword(password));
+        assertFalse(requirements.validPassword(password));
         password = "Bill1";
-        assertEquals(false, requirements.validPassword(password));
+        assertFalse(requirements.validPassword(password));
         password = "Bill$";
-        assertEquals(false, requirements.validPassword(password));
+        assertFalse(requirements.validPassword(password));
     }
 
     /**
@@ -39,10 +41,10 @@ public class UserRequirementsTest {
     public void validUsername() {
         UserRequirements requirements = new UserRequirements();
         String username = "Bill";
-        assertEquals(true, requirements.validUsername(username));
+        assertTrue(requirements.validUsername(username));
         username = "Bii";
-        assertEquals(false, requirements.validUsername(username));
+        assertFalse(requirements.validUsername(username));
         username = "B ill";
-        assertEquals(false, requirements.validUsername(username));
+        assertFalse(requirements.validUsername(username));
     }
 }
